cleanup

+++ simple-library-desktop/src/main/service/collection/collectionService.ts
@@ -1,6 +1,7 @@
 import { CollectionDataAccess } from '../../persistence/collectionDataAccess';
-import { Collection } from '../../../common/commonModels';
+import { ALL_ITEMS_COLLECTION_ID, Collection } from '../../../common/commonModels';
 import { ItemService } from "../item/ItemService";
+import { failAsync } from '../../../common/AsyncCommon';
 
 export class CollectionService {
 
@@ -22,14 +23,19 @@ export class CollectionService {
     public getAllCollections(includeItemCount: boolean): Promise<Collection[]> {
         return this.collectionDataAccess.getCollections(includeItemCount)
             .then(async (collections: Collection[]) => {
-                const allItemsCollection: Collection = {
-                    id: undefined,
-                    groupId: undefined,
-                    itemCount: includeItemCount ? await this.itemService.getTotalItemCount() : undefined,
-                    name: "All Items",
-                };
+                const allItemsCollection: Collection = await this.createAllItemsCollection(includeItemCount);
                 return [allItemsCollection, ...collections];
             });
+    }
+
+
+    private async createAllItemsCollection(includeItemCount: boolean): Promise<Collection> {
+        return {
+            id: ALL_ITEMS_COLLECTION_ID,
+            groupId: undefined,
+            itemCount: includeItemCount ? await this.itemService.getTotalItemCount() : undefined,
+            name: "All Items",
+        };
     }
 
 
@@ -50,7 +56,11 @@ export class CollectionService {
      * @return a promise that resolves when the collection was deleted
      */
     public deleteCollection(collectionId: number): Promise<void> {
-        return this.collectionDataAccess.deleteCollection(collectionId);
+        if (collectionId === ALL_ITEMS_COLLECTION_ID) {
+            return failAsync("Can not delete \"All-Items-Collection\".");
+        } else {
+            return this.collectionDataAccess.deleteCollection(collectionId);
+        }
     }
 
 
@@ -61,8 +71,13 @@ export class CollectionService {
      * @return a promise that resolves when the collection was renamed
      */
     public renameCollection(collectionId: number, newCollectionName: string): Promise<void> {
-        return this.collectionDataAccess.renameCollection(collectionId, newCollectionName);
+        if (collectionId === ALL_ITEMS_COLLECTION_ID) {
+            return failAsync("Can not rename \"All-Items-Collection\".");
+        } else {
+            return this.collectionDataAccess.renameCollection(collectionId, newCollectionName);
+        }
     }
+
 
     /**
      * Moves the collection with the given id into the group with the given id
@@ -71,8 +86,13 @@ export class CollectionService {
      * @return a promise that resolves when the collection was moved
      */
     public moveCollection(collectionId: number, targetGroupId: number | undefined): Promise<void> {
-        return this.collectionDataAccess.moveCollection(collectionId, targetGroupId ? targetGroupId : null);
+        if (collectionId === ALL_ITEMS_COLLECTION_ID) {
+            return failAsync("Can not move \"All-Items-Collection\".");
+        } else {
+            return this.collectionDataAccess.moveCollection(collectionId, targetGroupId ? targetGroupId : null);
+        }
     }
+
 
     /**
      * Move/Copy the given items to the given target collection
@@ -82,12 +102,32 @@ export class CollectionService {
      * @param copyMode whether to copy the items, i.e. not remove the items from the source collection
      */
     public async moveItemsToCollection(srcCollectionId: number, tgtCollectionId: number, itemIds: number[], copyMode: boolean): Promise<void> {
-        for (let i = 0; i < itemIds.length; i++) {
-            const itemId: number = itemIds[i];
-            if (copyMode) {
-                await this.collectionDataAccess.copyItemToCollection(tgtCollectionId, itemId);
-            } else {
-                await this.collectionDataAccess.moveItemsToCollection(srcCollectionId, tgtCollectionId, itemId);
+        if (tgtCollectionId === ALL_ITEMS_COLLECTION_ID) {
+            return failAsync("Can not move or copy items into \"All-Items-Collection\".");
+        } else {
+            for (let i = 0; i < itemIds.length; i++) {
+                const itemId: number = itemIds[i];
+                if (copyMode) {
+                    await this.collectionDataAccess.copyItemToCollection(tgtCollectionId, itemId);
+                } else {
+                    await this.collectionDataAccess.moveItemsToCollection(srcCollectionId, tgtCollectionId, itemId);
+                }
+            }
+        }
+    }
+
+
+    /**
+     * Removes the given items from the given collection
+     * @param collectionId the id of the collection
+     * @param itemIds the ids of the items to remove
+     */
+    public async removeItemsFromCollection(collectionId: number, itemIds: number[]): Promise<void> {
+        if (collectionId === ALL_ITEMS_COLLECTION_ID) {
+            return failAsync("Can not remove items from \"All-Items-Collection\".");
+        } else {
+            for (let i = 0; i < itemIds.length; i++) {
+                await this.collectionDataAccess.removeItemsFromCollection(collectionId, itemIds[i]);
             }
         }
     }
